Use a well-defined ABI and layout for `GlobalVariables`
The current definition of `GlobalVariables` implicitly relies on the ABI and layout of a copy-pasted definition of `GlobalVariables` to remain stable / the same across compilations, even across compilations using different compiler versions. **This is not guaranteed**, and is very likely to eventually lead to unsound breakage when that happens.

Since the only way to easily imbue a `struct` with a stable layout is the `#[repr(C)]` annotation, we use that to fix the outermost issue.

Alas, the issue then recurses onto the `String` and `PathBuf` definitions (_e.g._, image `PathBuf` being added a private extra field, or `String` having its `.ptr, .len, .cap` fields reordered): I am currently suggesting they be `Box`-wrapped so as to, at the very least, provide a stable layout to `GlobalVariables`. But since the dereferences of such pointers may still break in the future, ideally, the tools of a crate specialized on tackling these problems such as https://docs.rs/abi_stable ought to be used, but a decent simpler replacement could be the usage of well-defined types such as [`repr_c::String`](https://docs.rs/safer-ffi/0.0.5/safer_ffi/string/struct.String.html) (at the cost of replacing `PathBuf`s with `String`s, _i.e._, not supporting non-UTF8 paths).